backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic[email]==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
email-validator==2.1.0.post1
pytest==7.4.3
httpx==0.25.2
pytest-asyncio==0.21.1
python-multipart==0.0.20
python-dateutil==2.8.2
